dj_rag_0121/rag/requirements.txt
openai>=1.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-chroma>=0.1.0
transformers>=4.30.0
torch>=2.0.0
open_clip_torch>=2.20.0
